Enhance validation system and database handling for custom fields
- Update integer_value column type to bigInteger for better handling of large integers
- Introduce ValidationService for managing validation rules and constraints
- Implement SafeValueConverter for safe value conversions to prevent numeric overflow
- Add FieldTypeValidator for type-specific validation rules
- Improve validation logic to respect user-defined constraints and enforce database limits
- Add comprehensive tests for validation and safe value conversion

+++ database/migrations/create_custom_fields_table.php
@@ -124,7 +124,7 @@ return new class extends Migration
             $table->text('string_value')->nullable();
             $table->longText('text_value')->nullable();
             $table->boolean('boolean_value')->nullable();
-            $table->integer('integer_value')->nullable();
+            $table->bigInteger('integer_value')->nullable();
             $table->double('float_value')->nullable();
             $table->date('date_value')->nullable();
             $table->dateTime('datetime_value')->nullable();
